[hip-kernel-provider] Split up TestBatchnormApplicability.cpp
The hip-kernel-provider engine plan test for batch norm applicability is
unmaintainably large. Refactored this by splitting it up into the
different layers as defined by comments in the source. Generic helpers
have also been split into their own file header.

I also updated the directory structure so that all batchnorm plan tests
are in their own directory, as we add more plans this will help group
those tests.
