+++ dnn-providers/hip-kernel-provider/tests/engines/plans/Batchnorm/TestBatchnormFwdInferenceWithVariancePlan.cpp
