+++ dnn-providers/hip-kernel-provider/tests/engines/plans/Batchnorm/TestBatchnormPlanBuilder.cpp
