+++ dnn-providers/hip-kernel-provider/tests/engines/plans/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+# Copyright © Advanced Micro Devices, Inc., or its affiliates.
+# SPDX-License-Identifier:  MIT
+
+add_subdirectory(Batchnorm)
